+++ src/components/04-Components/navigation/navigation.js
@@ -118,6 +118,16 @@ class DesktopNavigation {
         }
       });
     }
+
+    for (const item of document.querySelectorAll('.masonry')) {
+      new Masonry( item, {
+        // options
+        itemSelector: '.masonry-item',
+        columnWidth: 320,
+        transitionDuration: 0,
+        containerStyle: { position: 'absolute' }
+      });
+    }
   }
 
   close() {